:recycle: [Refactor] 코드 정리
타입 재선언
컴포넌트 분리
테스트 코드 분리
