+++ src/api/moving-requests.ts
@@ -1,0 +1,23 @@
+import axiosInstance from "./axios";
+
+const PATH = "/moving-requests";
+
+export function createMovingRequest() {
+  /**
+   * - / ( POST )
+   * 1. Endpoint: `POST /moving-requests`
+   * 2. Description: 이사 요청 생성
+   * 3. Request : access-token 쿠키 전달
+   * 4. link : https://www.notion.so/API-14d9702f08878032932ee08ab2c19fb0?pvs=4#14d9702f0887814cb39efb0fa6c37733
+   */
+  return {};
+}
+export function getMovingRequestListByMover() {
+  /**
+   * 1. Endpoint:  `GET /moving-requests`
+   * 2. Description: 이사 요청 목록 조회
+   * 3. Request : access-token 쿠키 전달
+   * 4. link : https://www.notion.so/API-14d9702f08878032932ee08ab2c19fb0?pvs=4#14d9702f0887818e831dde53326a39e0
+   */
+  return {};
+}
