+++ src/types/mover.ts
@@ -42,6 +42,10 @@ export interface Address {
   dropOffAddress: string;
 }
 
+export interface ConfirmInfo {
+  isConfirmed: boolean;
+}
+
 export interface RequestDetails {
   movingDate: string;
   requestDate: string;
@@ -60,6 +64,38 @@ export interface QuoteDetailsData {
   dropOffAddress: string;
 }
 
-export interface ConfirmInfo {
-  isConfirmed?: boolean;
+interface UserBaseData {
+  id: number;
+  name: string;
+  email?: string;
+  phoneNumber?: string;
+}
+interface MoverBaseData extends UserBaseData {
+  imageUrl: string | null;
+  nickname: string;
+  career: number;
+  services: number[];
+  regions: number[];
+  introduction?: string;
+  description?: string;
+}
+// API 구조 참고
+export interface RatingData {
+  "1": number;
+  "2": number;
+  "3": number;
+  "4": number;
+  "5": number;
+  totalCount: number;
+  totalSum: number;
+  average: number;
+}
+// API 구조 참고
+export interface MoverDetailData extends MoverBaseData {
+  isDesignated: boolean;
+  isFavorite: boolean;
+  favoriteCount: number;
+  reviewCount: number;
+  confirmCount: number;
+  rating: RatingData;
 }
