+++ src/types/quote.ts
@@ -1,0 +1,5 @@
+export interface QuoteData {
+  id: number;
+  cost: number;
+  comment: string;
+}
